8.5. Рефакторинг. Переход на slice (offers & app state).
