+++ src/types.ts
@@ -1,5 +1,4 @@
-import { CITIES, OFFER_TYPES, AppRoute, AuthorizationStatus, LoadingStatus, SortType } from './consts';
-import { sortOffersByCity } from './utils';
+import { CITIES, OFFER_TYPES, AppRoute, AuthorizationStatus, SortType } from './consts';
 
 type Cities = typeof CITIES[number];
 
@@ -58,33 +57,4 @@ type SortingType = keyof typeof SortType;
 
 type Authorization = typeof AuthorizationStatus[keyof typeof AuthorizationStatus];
 
-type LoginData = {
-  email: string;
-  password: string;
-};
-
-type CommentData = {
-  comment: string;
-  rating: number;
-};
-
-type LoadedData<T> = {
-  data: T;
-  status: typeof LoadingStatus[keyof typeof LoadingStatus];
-}
-
-type AppState = {
-  offers: ReturnType<typeof sortOffersByCity>;
-  currentCity: Cities;
-  sortType: SortingType;
-  authorizationStatus: Authorization;
-  user: LoadedData<User | undefined>;
-  loadedOffers: LoadedData<Offer[]>;
-  loadedFullOffer: LoadedData<FullOffer | undefined>;
-  loadedFavoriteOffers: LoadedData<Offer[]>;
-  loadedNearOffers: LoadedData<Offer[]>;
-  loadedComments: LoadedData<Comment[]>;
-  isNewCommentLoading: boolean;
-}
-
-export type { Offer, FullOffer, Cities, Location, User, UserShort, Comment, Page, SortingType, Authorization, LoginData, CommentData, LoadedData, AppState };
+export type { Offer, FullOffer, Cities, Location, User, UserShort, Comment, Page, SortingType, Authorization };
